refactor: 맞춤법 검사 경로 수정
