+++ src/apps/server/resumes/services/resumes.service.ts
@@ -1,5 +1,3 @@
-import { ApiService } from '📚libs/modules/api/api.service';
-import { SpellCheckResult } from '📚libs/modules/api/api.type';
 import { ResumeRepository } from '📚libs/modules/database/repositories/resume.repository';
 import { Injectable, NotFoundException } from '@nestjs/common';
 import {
@@ -9,18 +7,13 @@ import {
 } from '🔥apps/server/resumes/dtos/get-resume.dto';
 import { PatchResumeRequestDto } from '🔥apps/server/resumes/dtos/patch-resume.dto';
 import { PostResumeResponseDto } from '🔥apps/server/resumes/dtos/post-resume.dto';
-import { PostSpellCheckRequestBodyDto } from '🔥apps/server/resumes/dtos/post-spell-check-request.body.dto';
 import { Question, Resume } from '@prisma/client';
 import { PrismaService } from '📚libs/modules/database/prisma.service';
 import { GetCountOfResumeResponseDto } from '🔥apps/server/resumes/dtos/get-count-of-resume.dto';
 
 @Injectable()
 export class ResumesService {
-  constructor(
-    private readonly resumesRepository: ResumeRepository,
-    private readonly apiService: ApiService,
-    private readonly prisma: PrismaService,
-  ) {}
+  constructor(private readonly resumesRepository: ResumeRepository, private readonly prisma: PrismaService) {}
 
   /**
    * 유저가 작성한 모든 자기소개서를 가져옵니다. 문항의 답안(answer)은 payload가 크기 때문에 option으로 선택해 가져옵니다.
@@ -110,13 +103,6 @@ export class ResumesService {
     // Entity -> DTO
     const resumeResponseDto = new PostResumeResponseDto(resume);
     return resumeResponseDto;
-  }
-
-  public async spellCheck(body: PostSpellCheckRequestBodyDto): Promise<SpellCheckResult[][]> {
-    const { sentence } = body;
-    const checkedSpellByDAUM = await this.apiService.spellCheckByDaum(sentence);
-
-    return checkedSpellByDAUM;
   }
 
   async deleteResume({ resumeId, userId }: { resumeId: number; userId: number }): Promise<void> {
